pyproject.toml
[tool.poetry]
name = "zkTLS-twitter-agent"
version = "0.1.0"
description = "An autonomous AI agent for zkTLS technology education"
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
tweepy = "^4.12.0"
python-dotenv = "^0.19.0"
requests = "^2.31.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
black = "^23.0.0"
mypy = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true 